// api/playauto_api_type.ts
export interface IPACategoryResponse {
    success: boolean
    result: IPACategoryArrayResult
}

export interface IPACategoryArrayResult {
    success: boolean
    makeDate: string
    total: number
    result: (IPACategoryResult | IPACategoryStoreResult)[]
}

export interface IPACategoryResult {
    number: number
    code: string
    c1: string
    c2: string
    c3: string
    c4: string
    c1_nm: string
    c2_nm: string
    c3_nm: string
    c4_nm: string
}
export interface IPACategoryStoreResult {
    number: number
    acode: string
    pcode: string
    ccode: string
    dc1: string
    dc2: string
    dc3: string
    dc4: string
    dc1_nm: string
    dc2_nm: string
    dc3_nm: string
    dc4_nm: string
    state: string
    cate_state_pdate: number | null
    cate_state_cdate: number | null
}